% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_todo.R
\name{find_todo}
\alias{find_todo}
\title{Will find and list all instances of the word "TODO" inside a project.}
\usage{
find_todo()
}
\description{
Will find and list all instances of the word "TODO" inside a project.
}
