% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_importfrom.R
\name{tidy_importfrom}
\alias{tidy_importfrom}
\title{tidy all importFrom statements}
\usage{
tidy_importfrom()
}
\description{
When this function is run then all `#' @importFrom package function`
statements, in the `R` in the current working directory, will be moved
to a file `R/0_imports.R`, cleaned and sorted.
}
\examples{
\dontrun{
tidy_importfrom()
}
}
